moved progress component to sumeets tab
