+++ Cyclops/src/app/page-space-su/progress-indicators/progress-indicators.component.ts
